requirement.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
configparser==5.2.0
cycler==0.11.0
docker-pycreds==0.4.0
docopt==0.6.2
dominate==2.6.0
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
imageio==2.16.1
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.0.0
pipreqs==0.4.11
promise==2.3
protobuf==3.19.3
psutil==5.9.0
pyparsing==3.0.7
python-box==5.4.1
python-dateutil==2.8.2
pytorch-msssim==0.2.1
pytz==2022.1
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.7.3
sentry-sdk==1.5.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
termcolor==1.1.0
tifffile==2022.2.9
torch==1.10.1
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
utils==1.0.1
visdom==0.1.8.9
wandb==0.12.9
websocket-client==1.2.3
yarg==0.1.9
yaspin==2.1.0
